+++ src/frontend/js/api/joanie.ts
@@ -15,7 +15,8 @@ import context from 'utils/context';
 import { JOANIE_API_VERSION } from 'settings';
 import { ResourcesQuery } from 'hooks/useResources';
 import { ObjectHelper } from 'utils/ObjectHelper';
-import { Maybe } from 'types/utils';
+import { Maybe, Nullable } from 'types/utils';
+import { PaymentSchedule } from 'types/Joanie';
 import { checkStatus, getFileFromResponse } from './utils';
 
 /*
@@ -97,6 +98,7 @@ export const getRoutes = () => {
           download: `${baseUrl}/orders/:id/invoice/`,
         },
         submit_for_signature: `${baseUrl}/orders/:id/submit_for_signature/`,
+        submit_installment_payment: `${baseUrl}/orders/:id/submit_installment_payment/`,
       },
       certificates: {
         download: `${baseUrl}/certificates/:id/download/`,
@@ -141,6 +143,9 @@ export const getRoutes = () => {
       },
       products: {
         get: `${baseUrl}/courses/:course_id/products/:id/`,
+        paymentSchedule: {
+          get: `${baseUrl}/courses/:course_id/products/:id/payment-schedule/`,
+        },
       },
       orders: {
         get: `${baseUrl}/courses/:course_id/orders/:id/`,
@@ -291,6 +296,11 @@ const API = (): Joanie.API => {
           fetchWithJWT(ROUTES.user.orders.submit_for_signature.replace(':id', id), {
             method: 'POST',
           }).then(checkStatus),
+        submit_installment_payment: async (id, payload) =>
+          fetchWithJWT(ROUTES.user.orders.submit_installment_payment.replace(':id', id), {
+            method: 'POST',
+            body: JSON.stringify(payload),
+          }).then(checkStatus),
       },
       enrollments: {
         create: async (payload) =>
@@ -417,6 +427,25 @@ const API = (): Joanie.API => {
           }
 
           return fetchWithJWT(buildApiUrl(ROUTES.courses.products.get, filters)).then(checkStatus);
+        },
+        paymentSchedule: {
+          get: async (
+            filters?: Joanie.CourseProductQueryFilters,
+          ): Promise<Nullable<PaymentSchedule>> => {
+            if (!filters) {
+              throw new Error(
+                'A course code and a product id are required to fetch a course product',
+              );
+            } else if (!filters.course_id) {
+              throw new Error('A course code is required to fetch a course product');
+            } else if (!filters.id) {
+              throw new Error('A product id is required to fetch a course product');
+            }
+
+            return fetchWithJWT(
+              buildApiUrl(ROUTES.courses.products.paymentSchedule.get, filters),
+            ).then(checkStatus);
+          },
         },
       },
       orders: {