+++ src/frontend/js/components/SaleTunnel/index.credential.spec.tsx
@@ -56,12 +56,6 @@ describe('SaleTunnel / Credential', () => {
     return <SaleTunnel {...props} isOpen={true} onClose={() => {}} />;
   };
 
-  const formatPrice = (price: number, currency: string) =>
-    new Intl.NumberFormat('en', {
-      currency,
-      style: 'currency',
-    }).format(price);
-
   setupJoanieSession();
 
   beforeEach(() => {
@@ -98,6 +92,10 @@ describe('SaleTunnel / Credential', () => {
         `https://joanie.endpoint/api/v1.0/orders/?course_code=${course.code}&product_id=${product.id}&state=pending&state=validated&state=submitted`,
         [],
       )
+      .get(
+        `https://joanie.endpoint/api/v1.0/courses/${course.code}/products/${product.id}/payment-schedule/`,
+        [],
+      )
       .post('https://joanie.endpoint/api/v1.0/orders/', (url, { body }) => {
         createOrderPayload = JSON.parse(body as any);
         return order;
@@ -120,7 +118,7 @@ describe('SaleTunnel / Credential', () => {
     await screen.findByText(getAddressLabel(billingAddress));
 
     const $button = screen.getByRole('button', {
-      name: `Pay ${formatPrice(product.price, product.price_currency)}`,
+      name: `Subscribe`,
     }) as HTMLButtonElement;
 
     const $terms = screen.getByLabelText(